More refactoring in progress
