+++ include/measurement_kit/common.hpp
@@ -11,13 +11,11 @@
 #include <measurement_kit/common/lexical_cast.hpp>
 #include <measurement_kit/common/logger.hpp>
 #include <measurement_kit/common/mock.hpp>
-#include <measurement_kit/common/net_test.hpp>
 #include <measurement_kit/common/non_copyable.hpp>
 #include <measurement_kit/common/non_movable.hpp>
 #include <measurement_kit/common/parallel.hpp>
 #include <measurement_kit/common/range.hpp>
 #include <measurement_kit/common/reactor.hpp>
-#include <measurement_kit/common/runner.hpp>
 #include <measurement_kit/common/settings.hpp>
 #include <measurement_kit/common/settings_entry.hpp>
 #include <measurement_kit/common/socket.hpp>
