+++ include/measurement_kit/nettests/ndt_test.hpp
@@ -1,15 +1,15 @@
 // Part of measurement-kit <https://measurement-kit.github.io/>.
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
-#ifndef MEASUREMENT_KIT_NETTESTS_NDT_NDT_TEST_HPP
-#define MEASUREMENT_KIT_NETTESTS_NDT_NDT_TEST_HPP
+#ifndef MEASUREMENT_KIT_NETTESTS_NDT_TEST_HPP
+#define MEASUREMENT_KIT_NETTESTS_NDT_TEST_HPP
 
-#include <measurement_kit/ooni.hpp>
+#include <measurement_kit/nettests/ooni_test.hpp>
 
 namespace mk {
 namespace nettests {
 
-class NdtTest : public ooni::OoniTest {
+class NdtTest : public OoniTest {
   public:
     NdtTest() : NdtTest(Settings()) {}
     NdtTest(Settings s);
