+++ include/measurement_kit/nettests/web_connectivity.hpp
@@ -1,16 +1,16 @@
 // Part of measurement-kit <https://measurement-kit.github.io/>.
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
-#ifndef MEASUREMENT_KIT_NETTESTS_OONI_WEB_CONNECTIVITY_HPP
-#define MEASUREMENT_KIT_NETTESTS_OONI_WEB_CONNECTIVITY_HPP
+#ifndef MEASUREMENT_KIT_NETTESTS_WEB_CONNECTIVITY_HPP
+#define MEASUREMENT_KIT_NETTESTS_WEB_CONNECTIVITY_HPP
 
+#include <measurement_kit/nettests/ooni_test.hpp>
 #include <measurement_kit/ooni.hpp>
-#include <measurement_kit/report.hpp>
 
 namespace mk {
 namespace nettests {
 
-class WebConnectivity : public ooni::OoniTest {
+class WebConnectivity : public OoniTest {
   public:
     WebConnectivity() : WebConnectivity("", {}) {}
