+++ include/measurement_kit/ooni.hpp
@@ -5,7 +5,6 @@
 #include <measurement_kit/ooni/error.hpp>
 #include <measurement_kit/ooni/nettests.hpp>
 #include <measurement_kit/ooni/ooni_reporter.hpp>
-#include <measurement_kit/ooni/ooni_test.hpp>
 #include <measurement_kit/ooni/templates.hpp>
 #include <measurement_kit/ooni/utils.hpp>
 #endif
