+++ src/libmeasurement_kit/ooni/ooni_reporter.cpp
@@ -2,12 +2,14 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
+#include <measurement_kit/nettests/ooni_test.hpp> /* XXX */
 #include <measurement_kit/ooni.hpp>
 
 namespace mk {
 namespace ooni {
 
 using namespace mk::report;
+using namespace mk::nettests;
 
 OoniReporter::OoniReporter(const OoniTest &parent) {
     settings = parent.options; // Copy the parent's settings
